requirements.txt
dj-database-url==0.4.1
Django==1.10.3
djangorestframework==3.5.3
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2.2
Faker==0.7.3
nose==1.3.7
coverage==4.2
coveralls==1.1
